+++ src/ytdl_sub/plugins/video_tags.py
@@ -1,3 +1,4 @@
+from typing import Any
 from typing import Dict
 
 from ytdl_sub.entries.entry import Entry
@@ -26,6 +27,15 @@ class VideoTagsOptions(PluginOptions):
     """
 
     _required_keys = {"tags"}
+
+    @classmethod
+    def partial_validate(cls, name: str, value: Any) -> None:
+        """
+        Partially validate video tags
+        """
+        if isinstance(value, dict):
+            value["tags"] = value.get("tags", {})
+        _ = cls(name, value)
 
     def __init__(self, name, value):
         super().__init__(name, value)
